% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_plot.R
\name{gof_plot}
\alias{gof_plot}
\title{Plot the Goodness-of-Fit Diagnostics}
\usage{
gof_plot(gof_stats, adjacency, degree_lim = 9, esp_lim = 8,
  geodesic_lim = 6)
}
\arguments{
\item{gof_stats}{as the Goodness-of-Fit statistics}

\item{adjacency}{as the observed adjacency matrix}

\item{degree_lim}{as a numeric with the limit of the degree which should be plotted}

\item{esp_lim}{as a numeric with the limit of the edgewise shared partner distribution which should be plotted}

\item{geodesic_lim}{as a numeric with the limit of the geodesic distance which should be plotted}
}
\value{
The plot with the degree distribution, the edgewise shared partner distribution and
minimum geodesic distance distribution of the simulated networks as boxplots with the observed statistics as red line
}
\description{
Plot the Goodness-of-Fit Diagnostics and the observed values of the network
}
\examples{
# generate symmetric matrix with diag = 0
N <- 10
set.seed(191919)
adjacency <- matrix(rbinom(n = N * N, size = 1, prob = 0.2), N, N)
diag(adjacency) <- 0
adjacency[lower.tri(adjacency)] <- t(adjacency)[lower.tri(adjacency)]
# generate smooth parameters
edge_param <- -2
## length 2star-vector: 2*(dim-1)-2+1  ## triangle: dim-2+1
star2_param <- c(0, -0.25, -0.5, -0.75, -1, -1.25, rep(-2, times = 2*(N-1)-2+1-6))
triangle_param <- c(0, -0.25, -0.5, -0.75, -1, -1.25, rep(-2, times = N-2+1-6))

# simulate networks and return Goodness-of-Fit statistics
gof_stats <- gof_pergm(dim = N, star2 = star2_param, triangle = triangle_param,
edge = edge_param, nsim = 2, method = "smoothERGM")
# plot the Goodness-of-Fit statistics
gof_plot(gof_stats, adjacency)

}

