% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_wrapper.R
\name{summary_stats_log}
\alias{summary_stats_log}
\title{Calculation of the log transformed statistics of edges, 2-star and triangle for an observed adjacency matrix}
\usage{
summary_stats_log(adjacency = NULL, n_cores = 4)
}
\arguments{
\item{adjacency}{The adjacency matrix of a network (symmetric with only zeros or ones)}

\item{n_cores}{The number of cores used for parallelization}
}
\value{
A vector of 3 with the number of edges, log transformed number of 2-stars and log transformed number of triangles
}
\description{
Calculates the log transformed statistics of edges, 2-star and triangle for an observed adjacency matrix. Calculation is combined in one loop in C++ and can be done in parallel.
}
\examples{
# generate symmetric matrix with diag = 0
N <- 10
set.seed(191919)
adjacency <- matrix(rbinom(n = N * N, size = 1, prob = 0.5), N, N)
diag(adjacency) <- 0
adjacency[lower.tri(adjacency)] <- t(adjacency)[lower.tri(adjacency)]

summary_stats_log(adjacency, n_cores = 4)

}
